Fix bug with HIP estimation and OnHeapCompressedFields
The issue was with the logic on the values being sent back to the callbacks.  The FieldsTest unit tests weren't verifying the values on the callback, once I added that verification, the bug became apparent.

I've also added a more "functional" test to the HipHllSketchTest while also adding the OnHeapCompressedFields to the list of implementations verified.