Comprehensive changes to quantiles, some Memory changes.
Fixed a subtle bug in quantiles that could have affected accuracy.
Refactored the code so that similar functions are in the same class.
Redesigned updates for Direct quantile sketches to be directly in
Memory. Direct Union operations remain on-heap for now.

Changed MemoryRegion to be immutable. The only place it was used was one
place in Tuple Sketches.

Improved testing and documention (in the test hierarchy) of how to use
the new MemoryUtil.memoryRequestHandler and examples of simple Memory
Managers that utilize the MemoryRequest interface.

Added a number of unit tests to improve coverage.

Improved documentation where I found ambiguous or incomplete language.