The use of "key" had multiple meanings.
In some places it meant the user supplied key and other places it meant
the hash of the user key. This was very confusing. This has been fixed.

Similarly "keys_" meant a compact array of hashes and in other places a
hash table. The arrays keys_ and summaries_ in Sketch<S> had overloaded
uses.  These two arrays were moved down a level to the CompactSketch and
the QuickSelectSketch to eliminate this confusion.

In the process of correcting this renaming and separation of uses, I
uncovered numerous questionable codings that could easily have caused
bugs. Certainly these confusions made the code harder to maintain and
understand.