Corrections based on Pavel's feedback:
ReqCompactor:

Implemented "ensureEnoughSections()" simlar to python although this
stops when the the new sectionSize would make it smaller than MIN_K.

The merge method calls ensureEnoughSections (once). It also chooses the
direction of the merge so that the smaller of the two merges into the
larger one.

In line 106 I wrote what I think you said on the call: "TODO if we are
at minK no need to compute these ??" Really?  I don't think this is
true. Even if we are at MIN_K, the compaction start and end will change
every time we compact.  I probably misunderstood what you meant :)

ReqSketch:

I eliminated the "lazy" option.