Code additions requested by Will.
Added 4 methods to Sketches class:

heapifyCompactSketch(Memory)
heapifyCompactSketch(Memory, long)
wrapCompactSketch(Memory)
wrapCompactSketch(Memory, long)

Changed the behavior of heapifySketch(2) and wrapSketch(2) calls
so that if the given image is a CompactSketch the behavior will
be the same as above.

The basic behavior change is that a seed is no longer required to
heapify or wrap a CompactSketch image.
But if it is given, it will be used to check the
hashSeed of the image, if there is one.

This push represents the main code changes.
More test code needs to be added.