I followed Jon's advice and made the various pairs of utility methods symmetric in their naming.
And, while I was at it, and since this will be a breaking change, I
improved the naming of a bunch of other methods as well.

And I added complementary methods where they were missing, such as
having similar tests for both int and longs.

I also added or improved the unit tests as well.