Interim 7. Final adjustments.
Added getQuantile{Upper/Lower}Bound to all the sketches and is now part
of the Quantiles{Floats/Doubles}APIs.

This means that a javadocs for the vast majority of all the methods are
now centralized into 3 API classes, which should make maintenance of the
javadocs much easier.

In parallel with this I spent a lot of time improving the wording and
clarity of the javadocs, especially those specified by the new
Quantiles*APIs.