Found a minor discrepancy between the documentation of the getQuantiles(int numEvenlySpaced) methods and the implementation of the QuantilesUtil.evenlySpaced(...) method.
The javadocs of the getQuantiles methods (for all quantiles sketches)
claimed that a "numEvenlySpaced == 1" was valid, yet the implementation
would throw an exception if "numEvenlySpaced < 2".

So I updated the implementation so that the implementation matched the
javadocs.

I also updated the relevant quantiles javadocs to reflect that this
exception could be thrown.

The only place this QuantilesUtil.evenlySpaced(...) method was used was
for these quantiles methods, so no place else in the library used this
method.