I discovered another bug in the design of the method getQuantiles(evenlySpaced).
The general method in QuantilesUtil.evenlySpacedFloats which is used in
several places in the library was also being used in the
getQuantiles(evenlySpaced) method.  It turns out that for the
getQuantiles, the evenlySpaced algo must omit 0.0, which the general
method does not.  It really needs its own generator for the evenly
spaced ranks.

Given that I fixed this issue, the documentation was also not correct,
so the javadocs had to be revised, again.