This set of changes fixed an "even space" bug.
But more than that it introduces a new dedicated method to assist in
creating partitions. This is the getPartitionBoundaries(int) method.

The return type is a small class that encapsulates all the information
that is required for a user to actually create partitions and load them
with data. From this class the user can not only populate the partitions
but also calculate their densities, weight and size prior to the actual
loading of the data into the partitions.