Mostly added tests
- The /quantilescommon/CrossCheckQuantilesTest lacked testing of the new
KllItemsSketch. This is a very comprehensive test of all of our quantile
sketches.

- The /quantilescommon/LinearRanksAndQuantiles lacked testing of the new
KllItemsSketch

- I had to move some utilities from the KllItems* classes to
/quantilescommon/Util so they could be leveraged by the quantilesCommon
test classes

- I discovered that the Kll*SerDeTests lacked a round-trip serde from
heap to off-heap and back. So I extended the tests for floats, doubles
and items.

Main code changes: very few.

- I removed some of the methods that are not used in the current
implementation of KllHeapItemsSketch and KllDirectCompactItemsSketch.

The code coverage is now pretty good and the KllItems classes have a
coverage > 95%.

- I found a method name spelling error in KllFloatsSketch that was
duplicated in KllDoublesSketch and referenced in other classes.

- removed temporary print methods from the KllDoublesSketch

This compiles with no errors in Eclipse and Maven.
Also the mvn clean javadoc:javadoc completes with no errors.
No checkstyle warnings.