Made some minor adjustments to the KLL toString(boolean, boolean) methods.  These changes make it easier to control what exactly is printed.
The toString() method in KllSketch should have been abstract. Changed
the parameter names to be consistent with what is actually implemented.

The Kll Items, Floats and Doubles Classes override the KllSketch
methods.  This is not new. I just changed the parameter names to match
the actual impl method in KllHelper.

In KllHelper I made printing the summary the default and it always
prints. In addition, the user can choose to print a small summary of the
Levels information, and /or the detail of the entire ItemsArray.

All of these options were there before. Just the parameters didn't
describe them very well.