Primary changes (so far) are to ItemsSketchSortedView::getPartitionBoundaries(..).
This change dramatically improved the handling of the end points and
isolates this adjustment from the SortedView class members.

The changes to SortedViewIterator and GenericSortedViewIterator are to
add some functionality of getting the key iteration variables without
having to specify the sorting criteria.

Added a specific test to check the upper and lower boundary adjustments
of a partition.