This consolidates the use of LS everywhere.
There were many copies of "public static final String LS =
System.getProperty("line.separator");" across many class files across
the repository. Although small, it just ends up using extra space in
each of the class files when it can be static imported from one place.
This fixes that.

Although it is a lot of files, it is the same fix in each one of them.