Corrections from Zac's code.
In the main code, nearly all the corrections are in the Javadocs.

However, in the test code, there were many places where he was comparing
a long vs a float ('cause he copied the code from the float classes).

Also, he didn't check the test classes with printing enabled. There were
numerous places where the printf format expected an "%f" that should
have been a "%d".

Nonetheless, Zac did an amazing job of creating this KllLongsSketch,
considering the complexity of the code!

I also did a coverage tests and made sure that the coverage of the
"longs" classes were very close to the coverage of the comparable
"doubles" and "floats" classes -- and they were!  This means that he was
able to faithfully translate all the relevant tests from float to longs
and I did not find any gaps in testing.