Lots of cleanup after Sketches class was removed.
Fixed some very bad assumptions with heapify(...) and wrap(...) where if
the simple call was made, i.e., heapify(seg) not only was the default
update seed assumed (this is normal), but the seed hash was not checked!
Yikes!  Same thing with wrap(...).  I'm not sure how that ever happened,
but it is now fixed.  I think a lot of that was associated with the old
SerVer 1 and 2 stuff.

This also fixed the squirrelly testing that somehow figured that the
above was ok!?!

Also added some checks when reading the preamble of a user's input
segment -- made sure that the segment was at least large enough to hold
the full preamble.